+++ lib/youtube_list.rb
@@ -11,18 +11,24 @@ require_relative '../lib/youtube_list/m3u8_interpreter'
 require_relative '../lib/youtube_list/raw_txt_interpreter'
 require_relative '../lib/youtube_list/youtube_player'
 
+# ocra execution preventer
 if(defined?(Ocra))
   exit()
 else
   begin
-    system('title YouTube Lister')
-    # for thesting purpose use File.absolute_path('./test') as argument of next method
-    Main.new()
-  rescue RuntimeError => e
-    warn(e.message)
-    warn(e.backtrace)
-    warn('An error has arised. Please take a screen-capture and send it to the programmer.')
-    sleep(60)
+    system('title Weightless Player')
+    Main.new(ARGV.first || File.absolute_path(File.dirname(ENV['OCRA_EXECUTABLE'])))
+  rescue StandardError, NoMemoryError, SystemStackError => e
+    ## SPECIAL CASE BEGIN
+    # I dont want to show errors to ppl
     exit!()
+    ## SPECIAL CASE END
+    warn('')
+    warn("Error: #{e.class}.")
+    warn("Message: #{e.message}.")
+    warn("Backtrace: #{e.backtrace.inspect}.")
+    warn('Please take a screen-capture and send it to the programmer.')
+    sleep(60)
+    exit()
   end
 end